# openapi.yaml
openapi: 3.0.1
info:
  title: Presentation Plugin
  description: A plugin that allows the user to create slides using ChatGPT. It allows you to say different things and it will return the right slides back. You should share the link to the slides so the user can download them
  version: 'v1'
servers:
  - url: http://localhost:80
paths:
  /presentation/link:
    post:
      operationId: getPresentationLink
      summary: Get the download link for the presentation file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addPresentationRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/presentationLinkResponse'

  /presentation/download/{presentation_id}:
    get:
      operationId: downloadPresentation
      summary: Download the presentation file using a unique identifier
      parameters:
        - name: presentation_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/pptx:
              schema:
                type: string
                format: binary
components:
  schemas:
    addPresentationRequest:
      type: object
      required:
        - slide_data
      properties:
        slide_data:
          type: array
          description: You need to pass in an object with slide_data then an array of objects in a json array with a heading (string), content (an array of points that will be mentioned on the slide)
          items:
            type: object
            required:
              - heading
              - content
            properties:
              heading:
                type: string
              content: 
                type: array
                items:
                  type: string
    presentationLinkResponse:
      type: object
      properties:
        download_link:
          type: string